QA #504: sanitize placements on the boot path too (not just install)
QA caught a real gap: the install path sanitizes (notifyInstalled ->
registerPlacements -> sanitizePlacements), but the boot path (runtime.ts
workspace loop) called placementRegistry.register directly with the RAW
instance.ui.placements. Since placements persist unfiltered on the BundleRef, a
spoof dropped at install would re-register verbatim on the next restart.

- Extract sanitizePlacements (+ PLACEMENT_STRING_MAX) from lifecycle.ts to
  defaults.ts (pure module next to hostMetaToUiMeta) so the boot path can import
  it without the lifecycle manager.
- runtime.ts boot loop now sanitizes before registering — same fail-closed guard
  as install.
- Tests: registry-level coverage of the registration contract BOTH paths share
  (sanitize -> register -> read-back); a mixed-authority spoof never reaches the
  registry, an all-malformed set registers nothing.

Note: QA's claim that fresh install needs a restart to surface the app is
incorrect — handleInstallRemoteOAuth calls notifyInstalled (registers sanitized
+ emits bundle.installed for live sidebar refresh). Verified.
